Implement prefix for ZFS dataset tree
Instead of using pool/iocage, we can now use anything like
pool/some/prefix/iocage, which enables iocage use within jails. To do
so, use the -p / --prefix option with iocage activate. The prefix is
then stored as a ZFS property of the pool and used for all commands.
A fixture is also added to the tests to enable functional tests.